requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
blas==1.0
blinker==1.4
brotlipy==0.7.0
ca-certificates==2022.07.19
cachetools==5.2.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
colorama==0.4.5
cryptography==37.0.1
cudatoolkit==11.3.1
dj-database-url==0.5.0
django==4.0.6
django-heroku==0.3.1
freetype==2.10.4
frozenlist==1.3.1
fsspec==2022.8.2
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.42.0
gunicorn==20.1.0
icc_rt==2019.0.0
idna==3.3
importlib-metadata==4.11.4
intel-openmp==2021.4.0
joblib==1.1.0
jpeg==9e
lerc==3.0
libdeflate==1.8
libpng==1.6.37
libprotobuf==3.20.1
libtiff==4.4.0
libuv==1.40.0
libwebp==1.2.2
lz4-c==1.9.3
markdown==3.4.1
markupsafe==2.1.1
mkl==2021.4.0
mkl-service==2.4.0
mkl_fft==1.3.1
mkl_random==1.2.2
multidict==6.0.2
numpy==1.23.1
numpy-base==1.22.3
oauthlib==3.2.1
openssl==1.1.1q
packaging==21.3
pillow==9.2.0
pip==21.2.4
protobuf==3.20.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
pydeprecate==0.3.2
pyjwt==2.4.0
pyopenssl==22.0.0
pyparsing==3.0.9
pysocks==1.7.1
python==3.9.12
python_abi==3.9
pytorch==1.12.1
pytorch-lightning==1.7.5
pytorch-mutex==1.0
pytz==2022.1
pyu2f==0.1.5
pyyaml==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.1
scipy==1.8.1
setuptools==63.2.0
six==1.16.0
sqlite==3.38.5
sqlparse==0.4.2
tensorboard==2.10.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tk==8.6.12
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.12.1+cpu
torchmetrics==0.9.3
torchvision==0.13.1+cpu
tqdm==4.64.1
typing-extensions==4.1.1
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.11
vc==14.2
vs2015_runtime==14.27.29016
werkzeug==2.2.2
wheel==0.37.1
whitenoise==6.2.0
win_inet_pton==1.1.0
wincertstore==0.2
xz==5.2.5
yaml==0.2.5
yarl==1.7.2
zipp==3.8.1
zlib==1.2.12
zstd==1.5.2
